Refactor and update solution
Reworked the solution using a different  technique. Instead of trying to have a for loop with logic to decide if it should add a ' ', '#' or '\n', I used nested for loops. The outer one keeps track of the line and whether we're on an even or odd line. The next for loop (depending on even or odd) determines which '#' or ' ' symbol to start with and alternates the symbols from there. Then tacks on a new line character