Create aList.js
Not close to being finished, but I finally realized how to create a list from an array. Instead of starting from the outermost object in the list and working my way in, I am starting with the innermost object (and setting it's rest to null) and then working my way out. It's must easier that way